backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
apturl==0.5.2
asgiref==3.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2025.11.12
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
distlib==0.4.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==3.0.1
Django==5.2.7
django-allauth==0.63.6
django-cloudinary-storage==0.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
duplicity==0.8.21
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fasteners==0.14.1
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.58.1
frozenlist==1.8.0
future==0.18.2
git-filter-repo==2.34.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.56.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.11
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.36.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.3
matplotlib==3.10.3
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
multidict==6.7.0
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==2.2.5
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
pillow==12.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.1.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.4.0
razorpay==2.0.0
reportlab==3.6.8
requests==2.32.5
rsa==4.9.1
seaborn==0.13.2
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
systemd-python==234
tenacity==9.1.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.8.8
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.2.0
urllib3==2.5.0
usb-creator==0.3.7
virtualenv==20.35.4
wadllib==1.3.6
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.4
whitenoise==6.11.0
xdg==5
xkit==0.0.0
yarl==1.22.0
zipp==1.0.0
